everything: Overlay, resizing, and rootfs building
Adds in base /etc files to the overlay, a sources.list file that matches
upstream Debian, and sets up a skeleton homedir for the enterprise user.

Adds in a resize_image.sh script to ensure the sdcard image has the
correct size to accommodate the rootfs.

Adjusts the Makefile to build the rootfs inside a directory first, then
rsync it to a generated sdcard image based upon the source image
provided to the Makefile.

Change-Id: Ia8c095141254bb114d8a9b3b6f82c19f5f3037f0