gpu: Add support for integrating vivante drivers
Adds support for syncing the Vivante binary Linux drivers into the rootfs
without packaging. We don't include the actual vivante sharfile because of their
restrictive license agreement. In the meantime, we can at least produce one-off
rootfses with the binaries so we can still continue developing.

This also introduces changes to the fix_permissions script to skip symlinks,
since setting permissions on symlinks is a meaningless operation.

Change-Id: Iba7dcbcd369647991739513dca38380da24ac4da