debootstrap: Build tarballs in /cache
This change allows us to build a debootstrap tarball containing debs that we
want to be present based upon DEBOOTSTRAP_EXTRA and keep it in a local cache
directory for quicker builds.

Note that if you update DEBOOTSTRAP_EXTRA, you'll want to run the
make-bootstrap-tarball again! The target is phony, and is not run by default to
reduce the number of round trips to Debian's upstream mirrors. At some point
we'll automate this to fetch tarballs straight from kokoro and manage the
package list in a different file somewhere else.

Change-Id: I09eef33b9f05c4ad5bf526e9c07dbdbf0157168c