kernel: Explicitly sequence steps needed to build
For some bizarre reason, the Kbuild doesn't like to be implicitly
sequenced in the way things are built. We need to do so on our end, so
we have to reduce some of our parallelization by explicitly sequencing
calls to each target rather than letting Make do stupid things and stomp
on itself.

Change-Id: Iba7076409694fcff7803eb95eee267042aa38ac7