Fix sdcard build
- After some of the refactoring that's happened, we ended up in a state
where the sparse rootfs img was being copied into the sdcard instead of
the non-sparse. Whoops.
- Also, update the boot script a little bit: Depending on the boot
device, we either use the PARTUUID of the system partition on MMC, or
the /dev path for the SD card. Unfortunately, the PARTUUID changes each
time the SD card image is generated, so we can't currently boot that by
UUID.

Change-Id: I66f132095adda79efaed6d6c0b7d387f33626754