Enable FETCH_PACKAGES by default
Should speed up most local builds. The theory is that most people
doesn't need to build all packages all the time and for those who
do or are working on a package builds with FETCH_PACKAGES=false.

Since this defaults local changes to packages to NOP add a short
notice every build.

Change-Id: Ib6b6233f23c052b5b88fc2ce309e5e5022f30ca7