Tune ROOTFS_PATCHED_IMG target
If the recipe fails for whatever reason one can not do further
incremental builds reliably. The output, ROOTFS_PATCHED_IMG, is
created at the top of the recipe. Next build it's current and in
a bad state. Instead start with ROOTFS_RAW_IMG and only create
ROOTFS_PATCHED_IMG when all is done.

Change-Id: I7e4cb95d3373cc20caacc15f1b0512112ce4ee76