Fix rootfs multistrap
- Okay, seems like there's enough subtle differences between a modern
system and the old Ubuntu on Kokoro, even though we're in a container.
Specifically, it seems like binfmt_support behaves a bit differently. To
fix the errors trying to run dash's preinst script (and presumably,
anything farther down as well), deploy the qemu binary into the chrooted
filesystem, and after the configuring of the system is finished, remove
it.

Change-Id: I79b2c12b035c28ad1a0bd9b06bb3ef430f3ec60d