Remove PACKAGES_FETCH_ROOT_DIRECTORY and friends
- Packages live in Apt now, so there's no reason to carry around this
legacy bit. It also is tough for external builds, as without a complete
packages.tgz to start with, you were unable to build.
- This changes the fallback logic for when a package is missing to
simply create the stamp file, and then the rootfs process will get the
package from apt. For this reason, there's no need to guard packages
with IS_EXTERNAL, either.

Change-Id: I490098f722cb5cf06b7373c8f26821b49e3b4de6