safe-abandon: Add in a new command to prevent ultimate sadness
The "repo abandon" command has a really bad habit of reverting changes across
the entire set of projects it's assigned to manage if you forget to specify a
project. Sadly, "repo start" used to be symmetrical, but now only affects the
current project.

This adds a "safe-abandon" command to our build setup so that we have an option
to prevent the ultimate sadness that occurs when you accidentally forget to
specify which project to abandon a branch in.

You're welcome. :(

Change-Id: I29519ffd1c315d492ebdc555a77712bfa31c596a