hack: Force generate_update_tarball.py to only build for arm64
This needs to be reverted once we adjust the pbuilder build environments to
build two more envs for each architecture we support. The root of the issue is
that Debian buster's APT can't resolve Build-Depends on cross targets, and
unfortunately, when your pbuilder env is setup to be arm64, trying to build an
armhf package there will ultimately fail because apt can't resolve the cross
build target.

Basically, "native" builds aren't currently "native" for armhf.

Change-Id: Id89ba8a065b785b0691e151307101859ddbe83fb