native-base: Fix the aarch64 qemu binaries in memory
On Ubuntu 18.04 just about everything is in place to allow qemu-user-static
interpreters to be used inside of other namespaces. Unfortunately, Ubuntu
18.04's qemu-user-static doesn't actually turn on the fix-binary flag for its
interpreters.

This patch force-enables the fix-binary flag for qemu-aarch64 so we can at least
create the native-base.tgz pbuilder environment. Once it's been built, we still
copy the interpreters into the environment so we don't have to force enable it
every single time we build.

Change-Id: I3ea9cb6f203f4a2989c7fcc3b51dbbff7af38ebe